% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.r
\docType{data}
\name{DengueSimR01}
\alias{DengueSimR01}
\title{Simulated dataset of dengue transmission with basic reproductive number of 1}
\format{
Matrix with five columns representing the X and Y coordinates of infected individuals, the time of infection,
the genotype of the infecting pathogen and the serotype of the infecting pathogen.
}
\usage{
DengueSimR01
}
\description{
Dataset simulated using an agent based model with a spatially heterogeneous population structure. Infectious agents
were introduced resulting in agent to agent transmission. The distance between successive cases in a transmission chain
were randomly drawn from a uniform distribution U(0,100). Each infectious agent resulted in a single transmission to
another agent after a delay of 15 days, reflecting the generation time of dengue. There are 11 transmission chains,
each with a different genotype. The genotypes are subdivided into four serotypes.
}
\author{
Justin Lessler and Henrik Salje
}
\keyword{datasets}
