% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{DengueSimulationR02}
\alias{DengueSimulationR02}
\title{Simulated dataset of dengue cases with basic reproductive number of 2}
\format{Matrix with five columns representing the X and Y coordinates of infected individuals, the time of infection, the genotype of the infecting pathogen and the serotype of the infecting pathogen.}
\usage{
DengueSimulationR01
}
\description{
Dataset simulated using an agent based model with a spatially heterogeneous population structure. Infectious agents were introduced resulting in agent to agent transmission. The distance between successive cases in a transmission chain were randomly drawn from a uniform distribution U(0,100). Each infectious agent resulted in transmissions to two other agents after a delay of 15 days, reflecting the generation time of dengue. There are 11 transmission chains, each with a different genotype. The genotypes are subdivided into four serotypes.
}
\author{
Justin Lessler and Henrik Salje
}

