% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transdistfuncs.r
\name{est.transdist.theta.weights}
\alias{est.transdist.theta.weights}
\title{Estimate transmission distance theta values by replication}
\usage{
est.transdist.theta.weights(case.times, gen.t.mean, t.density, t1,
  n.rep = 100)
}
\arguments{
\item{case.times}{a vector giving the occurrence time for each case}

\item{gen.t.mean}{the mean generation time of the infecting pathogen}

\item{t.density}{a vector giving the generation time density of the infecting pathogen}

\item{t1}{time step to begin simulation}

\item{n.rep}{number of replications in the simulation (default = 100)}
}
\value{
a three-dimensional array containing the mean normalized theta weights estimated across all replications
}
\description{
This function estimates the weight of each theta value by performing a user defined number of replications with the \code{get.transdist.theta} function. The weights
of each theta are calculated as the number of simulations in which a case at time \code{t1} and \code{t2} are separated by theta transmission events.
}
\references{
Salje H, Cummings DAT and Lessler J (2016). “Estimating infectious disease transmission distances using the overall distribution of cases.” Epidemics, 17, pp. 10–18. ISSN 1755-4365, doi: \href{https://www.sciencedirect.com/science/article/pii/S1755436516300317}{10.1016/j.epidem.2016.10.001}.
}
\seealso{
Other transdist: \code{\link{est.transdist.bootstrap.ci}},
  \code{\link{est.transdist.temporal.bootstrap.ci}},
  \code{\link{est.transdist.temporal}},
  \code{\link{est.transdist}},
  \code{\link{get.transdist.theta}}
}
\author{
Justin Lessler, Henrik Salje, and John Giles
}
\concept{transdist}
