% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.r
\name{get.pi.permute}
\alias{get.pi.permute}
\title{get the null distribution of the \code{get.pi} function}
\usage{
get.pi.permute(posmat, fun, r = 1, r.low = rep(0, length(r)), permutations)
}
\arguments{
\item{posmat}{a matrix with columns type, x and y}

\item{fun}{the function to evaluate}

\item{r}{the series of spatial distances we are interested in}

\item{r.low}{the low end of each range....0  by default}

\item{permutations}{the number of permute iterations}
}
\value{
pi values for all the distances we looked at
}
\description{
Does permutations to calculate the null distribution of get pi
if there were no spatial dependence. Randomly reassigns coordinates
to each observation permutations times
}
\examples{
\dontrun{
 R/examples/get_pi_permute.R
 }

}
\seealso{
Other get.pi: \code{\link{get.pi.bootstrap}},
  \code{\link{get.pi.ci}},
  \code{\link{get.pi.typed.bootstrap}},
  \code{\link{get.pi.typed.permute}},
  \code{\link{get.pi.typed}}, \code{\link{get.pi}}
}
