% Generated by roxygen2 (4.0.2): do not edit by hand
\name{get.pi.typed.permute}
\alias{get.pi.typed.permute}
\title{get the null distribution of the get.pi.typed function}
\usage{
get.pi.typed.permute(posmat, typeA = -1, typeB = -1, r = 1,
  r.low = rep(0, length(r)), permutations)
}
\arguments{
\item{posmat}{a matrix with columns type, x and y}

\item{typeA}{the "from" type that we are interested in, -1 is wildcard}

\item{typeB}{the "to" type that we are interested i, -1 is wildcard}

\item{r}{the series of spatial distances we are interested in}

\item{r.low}{the low end of each range....0  by default}

\item{permutations}{the number of permute iterations}
}
\value{
pi values for all the distances we looked at
}
\description{
Does permutations to calculate the null distribution of get pi
if there were no spatial dependence. Randomly reassigns coordinates
to each observation permutations times
}
\examples{
data(DengueSimulationR02)

r.max<-seq(20,1000,20)
r.min<-seq(0,980,20)

#Lets see if there's a difference in spatial dependence between those that occurred late versus early in the outbreak
type<-2-(DengueSimR02[,"time"]<75)
tmp<-cbind(DengueSimR02,type=type)

typed.pi<-get.pi.typed(tmp,typeA=1,typeB=2,r=r.max,r.low=r.min)
typed.pi.type.null<-get.pi.typed.permute(tmp,typeA=1,typeB=2,r=r.max,r.low=r.min,permutations=100)
}
\author{
Justin Lessler and Henrik Salje
}
\seealso{
Other get.pi: \code{\link{get.pi.bootstrap}};
  \code{\link{get.pi.ci}}; \code{\link{get.pi.permute}};
  \code{\link{get.pi.typed.bootstrap}};
  \code{\link{get.pi.typed}}; \code{\link{get.pi}}
}

