% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.r
\name{get.tau.GET}
\alias{get.tau.GET}
\title{Global hypothesis testing for the tau statistic}
\usage{
get.tau.GET(posmat, fun, r, r.low, permutations = 2500, comparison.type)
}
\arguments{
\item{posmat}{a matrix with columns x, y and any other named
columns needed by \code{fun}}

\item{fun}{a function that takes in two rows of posmat and returns:
\enumerate{
     \item for pairs included in the numerator (and the denominator for independent data)
     \item for pairs that should only be included in the denominator
     \item for pairs that should be ignored all together}
Note that names from \code{posmat} are not preserved in calls to
\code{fun}, so the columns of the matrix should be referenced numerically
so this is not available to fun}

\item{r}{the series of spatial distances (or their maximums) we are
interested in}

\item{r.low}{the low end of each range, 0 by default}

\item{permutations}{number of simulations of H_0. 2,500 is an optimal number according to Myllymäki et al. (2017).}

\item{comparison.type}{what type of points are included in the comparison set.
\itemize{
  \item "representative" if comparison set is representative of the underlying population
  \item "independent" if comparison set is cases/events coming from an indepedent process
}}
}
\value{
An object of class \code{tauGET} which can then be plotted using \code{plot.tau()} and an additional \code{tau} class object. The object consists of:
\itemize{
  \item r that inputted earlier
  \item obs the tau point estimate computed internally using \code{get.tau()}
  \item central the median estimate of all simulation curves that represent the null hypothesis. Comparing this to the tau=1 line indicates if it is reasonable to assume that H_0 was adequately simulated.
  \item lo the lower bound of the global envelope
  \item hi the upper bound of the global envelope
  \item tau.permute the entire record of simulations of H_0, to plot with the global envelope using \code{plot.tau()}.
}
}
\description{
Performs a graphical hypothesis test to assess the evidence against the null hypothesis (H_0: tau = 1 i.e. no spatiotemporal clustering nor inhibition). A global envelope test from the \code{GET} package is used to see if any part of the point estimate connected line is outside the lower or upper bounds of the global envelope. The global envelope is formed on the tau estimator acting on time-permuted data to simulate H_0. The global envelope test is of 'extreme rank type' i.e. minimum of pointwise ranks with 95\% significance level.
}
\section{Attributes}{

\itemize{
  \item p_interval represents a range rather than a single p-value to assess the evidence against H_0. Accessed using \code{attr(x,"p_interval")}.
}
}

\examples{
\donttest{
# Load data
r.max<-seq(20,1000,20)
r.min<-seq(0,980,20)
r.mid<-(r.max+r.min)/2
sero.type.func<-function(a,b,tlimit=20){
        if(a[5]==b[5]&(abs(a[3]-b[3])<=tlimit)){rc=1}
        else{rc=2}
        return(rc)
}

data(DengueSimRepresentative)
sero.type.rep.func<-function(a,b,tlimit=20){
     if(a[5]==1&b[5]==1&(abs(a[3]-b[3])<=tlimit)){rc=1}
     else{if(a[5]==1&b[5]==-999){rc=2}else{rc=3}}
     return(rc)
}

# get point estimate
Dengue.tau = get.tau(DengueSimRepresentative, sero.type.rep.func, r.max, r.min, "representative", 
                     data.frame = TRUE)

# perform graphical hypothesis test using a global envelope test
Dengue.GET = get.tau.GET(DengueSimRepresentative, sero.type.rep.func, r.max,r.min, 
                         permutations = 50, "representative")

#plot point estimate with global envelope and simulation of the null distribution
plot.tau(x = Dengue.tau, r.mid = TRUE, GET.res = Dengue.GET)
}
}
\seealso{
Other get.tau: 
\code{\link{get.tau.D.param.est}()},
\code{\link{get.tau.bootstrap}()},
\code{\link{get.tau.ci}()},
\code{\link{get.tau.permute}()},
\code{\link{get.tau.typed.bootstrap}()},
\code{\link{get.tau.typed.permute}()},
\code{\link{get.tau.typed}()},
\code{\link{get.tau}()},
\code{\link{plot.tau}()}

Other spatialtau: 
\code{\link{get.pi}()},
\code{\link{get.tau.D.param.est}()},
\code{\link{get.tau}()},
\code{\link{get.theta}()},
\code{\link{plot.tau}()}
}
\author{
Timothy M Pollington
}
\concept{get.tau}
\concept{spatialtau}
