% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.r
\name{get.theta.typed}
\alias{get.theta.typed}
\title{Optimized version of \code{get.theta} for typed data.}
\usage{
get.theta.typed(posmat, typeA = -1, typeB = -1, r = 1,
  r.low = rep(0, length(r)))
}
\arguments{
\item{posmat}{a matrix with columns type, x and y}

\item{typeA}{the "from" type that we are interested in, -1 is wildcard}

\item{typeB}{the "to" type that we are interested i, -1 is wildcard}

\item{r}{the series of spatial distances wer are interested in}

\item{r.low}{the low end of each range....0  by default}
}
\value{
theta values for all the distances we looked at
}
\description{
Version of the \code{get.theta} function that is optimized for statically typed data. That is
data where we are interested in the odds that points within some distance of points of
typeA are of typeB.
}
\seealso{
Other get.theta: \code{\link{get.theta.bootstrap}},
  \code{\link{get.theta.ci}},
  \code{\link{get.theta.permute}},
  \code{\link{get.theta.typed.bootstrap}},
  \code{\link{get.theta.typed.permute}},
  \code{\link{get.theta}}
}
\author{
Justin Lessler and Henrik Salje
}
\concept{get.theta}
