% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.r
\name{plot.tau}
\alias{plot.tau}
\title{Plotting the results from tau functions}
\usage{
\method{plot}{tau}(
  x,
  r.mid = TRUE,
  tausim = NULL,
  ptwise.CI = NULL,
  GET.res = NULL,
  d.param.est = NULL,
  ...
)
}
\arguments{
\item{x}{\code{tau} object; create using \code{get.tau(..., data.frame = TRUE)}. Required for all plots.}

\item{r.mid}{If \code{TRUE}(default) then for each point the x-coordinate of the midpoint of a distance band is plotted and if \code{FALSE} the endpoint of the distance band is plotted.}

\item{tausim}{the set of spatially-bootstrapped simulations of \code{taubstrap} class; use \code{get.tau.bootstrap()} to obtain this. Required for Estimation of the clustering range plot.}

\item{ptwise.CI}{the set of pointwise CIs of \code{tauCI} class; create using \code{get.tau(..., data.frame = TRUE)}. Optional for the diagnostic plot but should not be supplied for the other plots.}

\item{GET.res}{is a required object for the graphical hypothesis test plot but should not be supplied for the other plots. It is obtained from \code{get.tau.GET(..., data.frame = TRUE)}. It ensures that the user has performed a graphical hypothesis test first and has considered there is evidence against H_0, before deciding to estimate the clustering range.}

\item{d.param.est}{a required object for Estimating the clustering range plot from \code{get.tau.D.param(..., data.frame = TRUE)}, but should not be supplied for the other plots. A \code{taubstrap} object will also be necessary.}

\item{...}{other arguments which are standard for \code{plot()} for plot customisation}
}
\description{
Three types of plots:
\enumerate{
\item Diagnostic plot to indicate the structure or magnitude of spatiotemporal clustering. Requires \code{tau} object; \code{tauCI} object optional to draw pointwise CIs. This plot is only suitable for the purpose of a graphical hypothesis test in the situation that a specific distance band is selected prior to graph creation.
\item Graphical hypothesis test to assess the evidence against the null hypothesis (no spatiotemporal clustering nor inhibition). Requires \code{tau} and \code{tauGET} objects.
\item Estimation of the clustering range (the distribution of the places on the horizontal tau=1 line, where decreasing bootstrap simulations first intercept). Requires \code{tau}, \code{tauparamest} and \code{taubstrap} objects.
}
}
\seealso{
Other get.tau: 
\code{\link{get.tau.D.param.est}()},
\code{\link{get.tau.GET}()},
\code{\link{get.tau.bootstrap}()},
\code{\link{get.tau.ci}()},
\code{\link{get.tau.permute}()},
\code{\link{get.tau.typed.bootstrap}()},
\code{\link{get.tau.typed.permute}()},
\code{\link{get.tau.typed}()},
\code{\link{get.tau}()}

Other spatialtau: 
\code{\link{get.pi}()},
\code{\link{get.tau.D.param.est}()},
\code{\link{get.tau.GET}()},
\code{\link{get.tau}()},
\code{\link{get.theta}()}
}
\author{
Timothy M Pollington
}
\concept{get.tau}
\concept{spatialtau}
