% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transdistfuncs.r
\name{sim.plot}
\alias{sim.plot}
\title{Plot output of simulated epidemic}
\usage{
sim.plot(sim)
}
\arguments{
\item{sim}{a three-column matrix object produced by the \code{sim.epidemic} function}
}
\value{
A two-panel plotted object
}
\description{
A simple visualization function which plots the location of the index case and the spatial distribution of subsequent cases, 
and the epidemic curve showing the case count over time.
}
\author{
John Giles, Justin Lessler, and Henrik Salje
}
