% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.r
\name{get.tau.bootstrap}
\alias{get.tau.bootstrap}
\title{Bootstrap \code{get.tau} values.}
\usage{
get.tau.bootstrap(posmat, fun, r = 1, r.low = rep(0, length(r)), boot.iter,
  comparison.type = "representative")
}
\arguments{
\item{posmat}{a matrix appropriate for input to \code{get.tau}}

\item{fun}{a function appropriate as input to \code{get.pi}}

\item{r}{the series of spatial distances wer are interested in}

\item{r.low}{the low end of each range....0  by default}

\item{boot.iter}{the number of bootstrap iterations}

\item{comparison.type}{the comparison type to pass as input to \code{get.pi}}
}
\value{
tau values for all the distances we looked at
}
\description{
Runs \code{get.tau} on multiple bootstraps of the data. Is formulated
such that the relationship between points and themselves will not be
calculated
}
\examples{
\dontrun{
 R/examples/get_tau_bootstrap.R
 }

}
\seealso{
Other get.tau: \code{\link{get.tau.ci}},
  \code{\link{get.tau.permute}},
  \code{\link{get.tau.typed.bootstrap}},
  \code{\link{get.tau.typed.permute}},
  \code{\link{get.tau.typed}}, \code{\link{get.tau}}
}
\author{
Justin Lessler and Henrik Salje
}
