% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatialfuncs.r
\name{get.tau.D.param.est}
\alias{get.tau.D.param.est}
\title{Cluster range estimation using \code{get.tau.D.param.est}}
\usage{
get.tau.D.param.est(r, boot.iter, tausim, GETres = NULL, ...)
}
\arguments{
\item{r}{the series of spatial distances (or their maximums) we are
interested in}

\item{boot.iter}{number of spatial bootstraps}

\item{tausim}{the set of spatially-bootstrapped simulations. Has to be \code{taubstrap} class; use \code{get.tau.bootstrap()} to obtain this.}

\item{GETres}{is a required object and is obtained from a previous global hypothesis test using \code{get.tau.GET}. It ensures that the user has performed a graphical hypothesis test first and has considered there is evidence against H_0, before deciding to estimate the clustering range.}
}
\value{
An object of class \code{tauparamest} which can then be plotted using \code{plot.tau()}. The object consists of:
\itemize{
  \item envelope the distribution of clustering range estimates
}
}
\description{
Estimates the range of spatiotemporal clustering. It records the place on the horizontal tau=1 line where each spatially bootstrapped simulation touches. This distribution then represents an empirical distribution for the clustering range and a confidence interval can be computed.
}
\section{Attributes}{

\itemize{
  \item BCaCI the BCa CI for the distribution of clustering range estimates
}
}

\seealso{
Other get.tau: 
\code{\link{get.tau.GET}()},
\code{\link{get.tau.bootstrap}()},
\code{\link{get.tau.ci}()},
\code{\link{get.tau.permute}()},
\code{\link{get.tau.typed.bootstrap}()},
\code{\link{get.tau.typed.permute}()},
\code{\link{get.tau.typed}()},
\code{\link{get.tau}()},
\code{\link{plot.tau}()}

Other spatialtau: 
\code{\link{get.pi}()},
\code{\link{get.tau.GET}()},
\code{\link{get.tau}()},
\code{\link{get.theta}()},
\code{\link{plot.tau}()}
}
\author{
Timothy M Pollington
}
\concept{get.tau}
\concept{spatialtau}
